Controllers are now expected to have a 'mainWidget':
        mainWidget is the "default" Control of the
        Controller, that which fires when you press
        enter.

this is used in the tests to find the '_widget' that should be
"activated".

Added mainWidget to FooController, BarController, Grid and Search. All
tests that passed pass, and the one test that errored out now only
fails. This is progress.

Found a small bug in Grid where if you didn't have a row selected,
everything went to sh*t.

Removed all references to _widget from common code (except for a test
that really needs to know)

Removed tabs from all over the place